+++ Vector/Vector.cs
@@ -4,23 +4,23 @@ namespace Vector
 {
     public class Vector2
     {
-        public double x, y;
+        public decimal x, y;
         public Vector2()
         {
             x = 0;
             y = 0;
         }
-        public Vector2(double x, double y)
+        public Vector2(decimal x, decimal y)
         {
             this.x = x;
             this.y = y;
         }
-        public double Length()
+        public decimal Length()
         {
-            return Math.Sqrt(x * x + y * y);
+            return (decimal)Math.Sqrt((double)(x * x + y * y));
         }
 
-        public static double operator *(Vector2 v, Vector2 w)
+        public static decimal operator *(Vector2 v, Vector2 w)
         {
             return v.x * w.x + v.y * w.y;
         }
@@ -35,12 +35,12 @@ namespace Vector
             return new Vector2(v.x - w.x, v.y - w.y);
         }
 
-        public static Vector2 operator *(Vector2 v, double x)
+        public static Vector2 operator *(Vector2 v, decimal x)
         {
             return new Vector2(v.x * x, v.y * x);
         }
 
-        public static Vector2 operator /(Vector2 v, double x)
+        public static Vector2 operator /(Vector2 v, decimal x)
         {
             return new Vector2(v.x / x, v.y / x);
         }
@@ -55,9 +55,10 @@ namespace Vector
         {
             return new Vector2(y, -x);
         }
+
         public static Vector2 Normalize(Vector2 v)
         {
-            double l = v.Length();
+            decimal l = v.Length();
             Vector2 n = new Vector2(v.x / l, v.y / l);
             return n;
         }
@@ -65,14 +66,14 @@ namespace Vector
 
     public class Vector3
     {
-        public double x, y, z;
+        public decimal x, y, z;
         public Vector3()
         {
             x = 0;
             y = 0;
             z = 0;
         }
-        public Vector3(double x, double y, double z)
+        public Vector3(decimal x, decimal y, decimal z)
         {
             this.x = x;
             this.y = y;
@@ -86,7 +87,7 @@ namespace Vector
             this.z = v.z;
         }
 
-        public static double operator *(Vector3 v, Vector3 w)
+        public static decimal operator *(Vector3 v, Vector3 w)
         {
             return v.x * w.x + v.y * w.y + v.z * w.z;
         }
@@ -101,13 +102,13 @@ namespace Vector
             return new Vector3(v.x - w.x, v.y - w.y, v.z - w.z);
         }
 
-        public double Length()
+        public decimal Length()
         {
-            return Math.Sqrt(x * x + y * y + z * z);
+            return (decimal)(Math.Sqrt((double)(x * x + y * y + z * z)));
         }
         public static Vector3 Normalize(Vector3 v)
         {
-            double l = v.Length();
+            decimal l = v.Length();
             Vector3 n = new Vector3(v.x / l, v.y / l, v.z / l);
             return n;
         }
